requirements.txt
amqp==5.1.1
APScheduler==3.6.3
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.6
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
decorator==5.1.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0
django-celery-beat==2.2.1
django-cors-headers==3.12.0
django-debug-toolbar==3.4.0
django-timezone-field==4.0
executing==0.8.3
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
ipython==8.3.0
jedi==0.18.1
kombu==5.2.4
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.11
pytube==12.1.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tornado==6.1
traitlets==5.2.1.post0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uvicorn==0.17.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.1.0
wrapt==1.14.1
